funcionalidad de editar un registro de producto

+++ src/main/java/com/practice/ecommerce/service/ProductService.java
@@ -7,8 +7,8 @@ import com.practice.ecommerce.model.Product;
 
 public interface ProductService {
 	public Product save(Product product);
-	public Optional <Product> get(Long id);
+	public Optional <Product> get(Integer id);
 	public void update(Product product);
-	public void delete(Long id);
+	public void delete(Integer id);
 	public List<Product> findAll();
 }
