+++ src/main/java/com/practice/ecommerce/service/ProductServiceImpl.java
@@ -21,7 +21,7 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	@Override
-	public Optional<Product> get(Long id) {
+	public Optional<Product> get(Integer id) {
 		return productRepository.findById(id);
 	}
 
@@ -32,7 +32,7 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	@Override
-	public void delete(Long id) {
+	public void delete(Integer id) {
 		productRepository.deleteById(id);
 		
 	}
